pyproject.toml
# https://pip.pypa.io/en/stable/reference/build-system/pyproject-toml/
[build-system]
# PEP 518 https://www.python.org/dev/peps/pep-0518/
# https://discuss.python.org/t/help-testing-experimental-features-in-setuptools/13821
requires = [
  "setuptools>=64",
  "wheel",               # for bdist package distribution
  "setuptools_scm>=6.4", # for automated versioning
  "libclang",
  "nanobind",
  "pyyaml",
]
# PEP 517 https://peps.python.org/pep-0517/
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
# include = []
# exclude = ["tests"]
namespaces = true

[tool.setuptools_scm]
write_to = "src/nanobindGenerator/_version.py"

# PEP 621 https://peps.python.org/pep-0621/
[project]
name = "nanobindGenerator"
# version = "0.0.1"
# for automated versioning by pypa/setuptools_scm. see https://peps.python.org/pep-0621/#dynamic
dynamic = ["version"]
description = "A tool for generating nanobind C++ bindings"
authors = [{ name = "Your Name" }, { email = "your.email@example.com" }]
maintainers = [{ name = "Your Name" }, { email = "your.email@example.com" }]
readme = "README.md"
license = { file = "LICENSE.txt" }
requires-python = ">=3.8"
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.12",
]

keywords = ["nanobind", "bindings", "generator", "cpp"]
dependencies = [
  "libclang",
  "nanobind",
  "pyyaml",
]

[project.scripts]
nbGen = "nanobindGenerator.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/nanobindGenerator"
documentation = "https://github.com/yourusername/nanobindGenerator"
repository = "https://github.com/yourusername/nanobindGenerator"
changelog = "https://github.com/yourusername/nanobindGenerator/blob/main/CHANGELOG.md"
